batch/arkbatchfunc/requirements.txt
azure-functions
azure-identity
azure-batch
msrest
